more bugfixes

+++ install/lib/Installer.php
@@ -235,30 +235,44 @@ class Installer
      * Stage 11 - Final completion message and clean up actions.
      */
     public function stage11(): void
-    {
+{
         global $session, $logd_version, $recommended_modules, $noinstallnavs, $stage, $DB_USEDATACACHE;
         output("`@`c`bAll Done!`b`c");
         output("Your install of Legend of the Green Dragon has been completed!`n");
         output("`nRemember us when you have hundreds of users on your server, enjoying the game.");
         output("Eric, JT, and a lot of others put a lot of work into this world, so please don't disrespect that by violating the license.`n`n");
         if ($session['user']['loggedin']){
-        	addnav("Continue",$session['user']['restorepage']);
+            addnav("Continue",$session['user']['restorepage']);
         }else{
-        	addnav("Login Screen","./");
+            addnav("Login Screen","./");
         }
         savesetting("installer_version",$logd_version);
         $file = __DIR__ . '/../../installer.php';
-                if (file_exists($file)) {
-                        try {
-                                if (unlink($file)) {
-                                        output("`\$Installer file installer.php removed.`n");
-                                } else {
-                                        output("`\$Unable to delete installer.php. Please remove it manually.`n");
-                                }
-                        } catch (Throwable $e) {
-                                output("`\$Error deleting installer.php: " . $e->getMessage() . "`n");
-                        }
+
+        // Check if the button has been pressed to delete the installer file
+        if (array_key_exists('delete_installer', $_POST) && $_POST['delete_installer'] == '1') {
+            if (file_exists($file)) {
+                try {
+                    if (unlink($file)) {
+                        output("`\$Installer file installer.php removed.`n");
+                    } else {
+                        output("`\$Unable to delete installer.php. Please remove it manually.`n");
+                    }
+                } catch (Throwable $e) {
+                    output("`\$Error deleting installer.php: " . $e->getMessage() . "`n");
                 }
+            }
+        } else {
+            if (file_exists($file)) {
+                output("`\$The installer.php file still exists. For security reasons, you should delete it now.`n");
+                rawoutput("
+                    <form method='POST' action='installer.php?stage=11'>
+                        <input type='hidden' name='delete_installer' value='1'>
+                        <input type='submit' class='button' value='".htmlentities(translate_inline("Delete installer.php now"), ENT_COMPAT, getsetting("charset", "ISO-8859-1"))."'>
+                    </form>
+                ");
+            }
+        }
         $this->checkDbconnectPermissions();
         $noinstallnavs=true;
     }
